+++ kitsutools/netlib_server.h
@@ -56,7 +56,7 @@ namespace ktools
 	{	USER_COMMAND_REQUEST
 	,	USER_COMMAND_RESPONSE
 	};
-	int32_t sendUserCommand(CClient* pClient, USER_COMMAND requestOrResponse, const char* buffer, size_t bufferSize);
+	int32_t sendUserCommand(CClient* pClient, USER_COMMAND requestOrResponse, const char* buffer, uint32_t bufferSize);
 };
 
 // Define this function for processing your commands.
